Cambios en LastFMServiceImpl
