+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/controller/MoreDetailsControllerInjector.kt
@@ -3,7 +3,6 @@ package ayds.lisboa.songinfo.moredetails.controller
 import ayds.lisboa.songinfo.moredetails.model.MoreDetailsModelInjector
 import ayds.lisboa.songinfo.moredetails.view.MoreDetailsView
 
-
 object MoreDetailsControllerInjector {
 
     fun onViewStarted(moreDetailsView: MoreDetailsView) =
